A3 Graph buildout
Undirected graph build out. I removed the Matrix and Adjacency List and Linked Objects classes. I made a graph class of linked objects and added display methods to output the data in the graphs in a Matrix and Adjacency List. I added a searching function to look-up a vertex in the graph. I changed insert to take in strings, not vertices. Working on Parse and Build files.