A3 Finished, main.pdf in build
Finished the whole LaTeX write-up. Literally at the last minute. But I got it all in and am happy with it. I made a folder for the build files and moved the .tex file to the outer folder. THE PDF IS IN THE BUILD FOLDER. Also finishing touches on files and stuff.