A4 SSSP Buildout
I allowed the Build file to record the first vertex inserted to be used later. I updated useGraph file to account for the SSSP. I changed main to implement the SSSP and report accurate info. I modified DirectedGraph adding the necessary recorders for the distance and predecessors for a vertex. I also made a private function to get the path and a public one to perform the Bellman-Ford algorithm over the graph from the starting vertex from Build. I also adjusted the formatting to make it easier to read.